# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ai/embeddings": {
      "post": {
        "description": "Encode un texte en vecteur sémantique avec OpenAI",
        "operationId": "AiController_generateEmbedding",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateEmbeddingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Embedding généré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Erreur lors de la génération de l'embedding"
          }
        },
        "summary": "Générer un embedding vectoriel",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/search": {
      "post": {
        "description": "Recherche dans la base de connaissances par similarité cosinus",
        "operationId": "AiController_semanticSearch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemanticSearchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Résultats de recherche",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticSearchResponseDto"
                }
              }
            }
          }
        },
        "summary": "Recherche sémantique",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/chat": {
      "post": {
        "description": "Assistant commercial basé sur RAG pour l'e-commerce",
        "operationId": "AiController_chatbot",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatbotQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Réponse du chatbot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatbotResponseDto"
                }
              }
            }
          }
        },
        "summary": "Chatbot conversationnel",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/chat/session": {
      "post": {
        "operationId": "AiController_createChatSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session créée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Créer une session de chat",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/recommendations": {
      "post": {
        "description": "Basées sur lhistorique dachat et le comportement",
        "operationId": "AiController_getRecommendations",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produits recommandés",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRecommendationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Recommandations de produits",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/sentiment": {
      "post": {
        "description": "Analyse le sentiment dun texte (avis, commentaire, etc.)",
        "operationId": "AiController_analyzeSentiment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SentimentAnalysisDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Résultat de lanalyse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SentimentAnalysisResponseDto"
                }
              }
            }
          }
        },
        "summary": "Analyse de sentiment",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/seo": {
      "post": {
        "description": "Génère des méta-données optimisées pour le SEO",
        "operationId": "AiController_generateSEO",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SEOGenerationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Optimisation SEO générée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SEOGenerationResponseDto"
                }
              }
            }
          }
        },
        "summary": "Génération SEO",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/knowledge": {
      "post": {
        "description": "Ajoute un document à la base de connaissances RAG",
        "operationId": "AiController_addKnowledgeDocument",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddKnowledgeDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document ajouté",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeDocumentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Ajouter un document (Admin)",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/knowledge/product": {
      "post": {
        "description": "Crée un embedding pour un produit",
        "operationId": "AiController_indexProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produit indexé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeDocumentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Indexer un produit (Admin)",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/ai/products/similar": {
      "get": {
        "description": "Recherche des produits similaires à une requête textuelle (ex: \"chaussures pour homme\") en utilisant la recherche vectorielle pgvector",
        "operationId": "AiController_searchSimilarProducts",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "description": "Requête de recherche",
            "schema": {
              "example": "chaussures pour homme",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre maximum de résultats",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Prix minimum",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Prix maximum",
            "schema": {
              "example": 500,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produits similaires trouvés",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": {
                      "type": "string"
                    },
                    "totalResults": {
                      "type": "number"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "productId": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "imageUrl": {
                            "type": "string"
                          },
                          "categoryName": {
                            "type": "string"
                          },
                          "similarityScore": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Rechercher des produits similaires",
        "tags": [
          "ai"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "API documentation for the application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "JWT-refresh": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Refresh-JWT",
        "description": "Enter refresh JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "GenerateEmbeddingDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Texte à encoder en embedding",
            "example": "Samsung Galaxy S24 Ultra smartphone Android 5G"
          },
          "documentId": {
            "type": "string",
            "description": "ID optionnel pour associer à un document",
            "example": "product-uuid-123"
          },
          "documentType": {
            "type": "string",
            "description": "Type de document",
            "enum": [
              "PRODUCT",
              "CATEGORY",
              "FAQ",
              "POLICY",
              "REVIEW",
              "GENERAL"
            ],
            "example": "PRODUCT"
          }
        },
        "required": [
          "text"
        ]
      },
      "EmbeddingResponseDto": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "description": "ID du document si fourni",
            "example": "product-uuid-123"
          },
          "embedding": {
            "type": "string",
            "description": "Embedding vectoriel encodé en base64",
            "example": "W3sidmFsdWUiOjAuMTIzLCJkZWYiOjB9XQ=="
          },
          "dimension": {
            "type": "number",
            "description": "Dimension du vecteur",
            "example": 1536
          },
          "model": {
            "type": "string",
            "description": "Modèle utilisé",
            "example": "text-embedding-ada-002"
          },
          "usage": {
            "type": "object",
            "description": "Tokens utilisés",
            "example": 10
          }
        },
        "required": [
          "documentId",
          "embedding",
          "dimension",
          "model",
          "usage"
        ]
      },
      "SemanticSearchDto": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "description": "Requête de recherche",
            "example": "smartphone bon appareil photo"
          },
          "documentTypes": {
            "type": "array",
            "description": "Types de documents à rechercher",
            "example": [
              "PRODUCT"
            ],
            "items": {
              "type": "string",
              "enum": [
                "PRODUCT",
                "CATEGORY",
                "FAQ",
                "POLICY",
                "REVIEW",
                "GENERAL"
              ]
            }
          },
          "limit": {
            "type": "number",
            "description": "Nombre maximum de résultats",
            "example": 10
          },
          "similarityThreshold": {
            "type": "number",
            "description": "Seuil de similarité minimum (0-1)",
            "example": 0.5
          }
        },
        "required": [
          "query"
        ]
      },
...
      "SemanticSearchResponseDto": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "description": "Requête utilisée",
            "example": "smartphone bon appareil photo"
          },
          "totalResults": {
            "type": "number",
            "description": "Nombre de résultats",
            "example": 10
          },
          "searchTimeMs": {
            "type": "number",
            "description": "Temps de recherche en ms",
            "example": 150
          },
          "results": {
            "description": "Résultats de recherche",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResultDto"
            }
          },
          "model": {
            "type": "string",
            "description": "Modèle utilisé",
            "example": "text-embedding-ada-002"
          }
        },
        "required": [
          "query",
          "totalResults",
          "searchTimeMs",
          "results",
          "model"
        ]
      },
      "ChatbotQueryDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message de l'utilisateur",
            "example": "Quel est le meilleur smartphone pour la photo?"
          },
          "sessionId": {
            "type": "string",
            "description": "ID de session de conversation existante",
            "example": "session-uuid-123"
          },
          "includeHistory": {
            "type": "boolean",
            "description": "Inclure l'historique récent dans le contexte",
            "example": true
          },
          "temperature": {
            "type": "number",
            "description": "Température pour la génération (0-1)",
            "example": 0.7
          }
        },
        "required": [
          "message"
        ]
      },
      "ChatbotResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "ID de la session",
            "example": "session-uuid-123"
          },
          "answer": {
            "type": "string",
            "description": "Réponse générée",
            "example": "Basé sur votre recherche, je vous recommande le Samsung Galaxy S24 Ultra..."
          },
          "sources": {
            "description": "Sources utilisées pour générer la réponse",
            "example": [
              "product-uuid-1",
              "product-uuid-2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "processingTimeMs": {
            "type": "number",
            "description": "Temps de génération en ms",
            "example": 2500
          },
          "model": {
            "type": "string",
            "description": "Nom du modèle utilisé",
            "example": "gpt-3.5-turbo"
          },
          "confidenceScore": {
            "type": "number",
            "description": "Score de confiance (0-1)",
            "example": 0.92
          },
          "recommendedProducts": {
            "description": "Produits recommandés",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "sessionId",
          "answer",
          "sources",
          "processingTimeMs",
          "model"
        ]
      },
      "CreateChatSessionDto": {
        "type": "object",
        "properties": {
          "metadata": {
            "type": "object",
            "description": "Métadonnées initiales de la session",
            "example": {
              "source": "website"
            }
          }
        }
      },
      "ChatSessionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de la session",
            "example": "session-uuid-123"
          },
          "userId": {
            "type": "string",
            "description": "ID de l'utilisateur",
            "example": "user-uuid-123"
          },
          "messages": {
            "description": "Messages de la conversation",
            "example": [
              {
                "role": "user",
                "content": "Bonjour",
                "timestamp": "2024-01-01T10:00:00Z"
              },
              {
                "role": "assistant",
                "content": "Bonjour! Comment puis-je vous aider?",
                "timestamp": "2024-01-01T10:00:01Z"
              }
            ],
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "createdAt": {
            "type": "string",
            "description": "Date de création",
            "example": "2024-01-01T10:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Date de dernière mise à jour",
            "example": "2024-01-01T10:30:00Z"
          }
        },
        "required": [
          "id",
          "messages",
          "createdAt",
          "updatedAt"
        ]
      },
      "RecommendationConfigDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "ID de l'utilisateur (injecté depuis le token si authentifié)",
            "example": "user-uuid-123"
          },
          "limit": {
            "type": "number",
            "description": "Nombre de recommandations",
            "example": 5
          },
          "preferredCategories": {
            "description": "Catégories à privilégier",
            "example": [
              "category-uuid-1",
              "category-uuid-2"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "minPrice": {
            "type": "number",
            "description": "Prix minimum",
            "example": 100
          },
          "maxPrice": {
            "type": "number",
            "description": "Prix maximum",
            "example": 1000
          }
        }
      },
      "ProductRecommendationResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "ID de l'utilisateur",
            "example": "user-uuid-123"
          },
          "recommendationType": {
            "type": "string",
            "description": "Type de recommandation",
            "example": "similar-to-purchased"
          },
          "products": {
            "description": "Produits recommandés",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "totalRecommendations": {
            "type": "number",
            "description": "Nombre de produits recommandés",
            "example": 5
          },
          "metadata": {
            "type": "object",
            "description": "Métadonnées de recommandation",
            "example": {
              "basedOnSearches": 3,
              "basedOnPurchases": 5,
              "avgSentiment": 0.75
            }
          }
        },
        "required": [
          "userId",
          "recommendationType",
          "products",
          "totalRecommendations"
        ]
      },
      "SentimentAnalysisDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Texte à analyser",
            "example": "Excellent produit, je recommande vivement!"
          },
          "documentType": {
            "type": "string",
            "description": "Type de texte",
            "enum": [
              "PRODUCT",
              "CATEGORY",
              "FAQ",
              "POLICY",
              "REVIEW",
              "GENERAL"
            ],
            "example": "REVIEW"
          }
        },
        "required": [
          "text"
        ]
      },
      "SentimentAnalysisResponseDto": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number",
            "description": "Score de sentiment (-1 à 1)",
            "example": 0.75
          },
          "label": {
            "type": "string",
            "description": "Label du sentiment",
            "enum": [
              "positive",
              "neutral",
              "negative"
            ],
            "example": "positive"
          },
          "confidence": {
            "type": "number",
            "description": "Confiance de l'analyse (0-1)",
            "example": 0.88
          },
          "positiveKeywords": {
            "description": "Mots-clés positifs détectés",
            "example": [
              "excellent",
              "recommande",
              "qualité"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "negativeKeywords": {
            "description": "Mots-clés négatifs détectés",
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "score",
          "label",
          "confidence",
          "positiveKeywords",
          "negativeKeywords"
        ]
      },
...
      "AddKnowledgeDocumentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Contenu textuel du document",
            "example": "Notre politique de retour permet de retourner les produits dans les 30 jours"
          },
          "documentType": {
            "type": "string",
            "description": "Type de document",
            "enum": [
              "PRODUCT",
              "CATEGORY",
              "FAQ",
              "POLICY",
              "REVIEW",
              "GENERAL"
            ],
            "example": "POLICY"
          },
          "title": {
            "type": "string",
            "description": "Titre du document",
            "example": "Politique de retour"
          },
          "metadata": {
            "type": "object",
            "description": "Métadonnées additionnelles",
            "example": {
              "section": "retour",
              "updatedAt": "2024-01-01"
            }
          },
          "externalId": {
            "type": "string",
            "description": "ID de référence externe",
            "example": "policy-return-v1"
          }
        },
        "required": [
          "content",
          "documentType"
        ]
      },
...
    }
  }
}